refactor(platform): PROTO-FREEZE-001 type permission contract
Tests: Qt 6.7.3 MinGW probe CTest 9/9 PASS; python product/scripts/validate-package.py PASS (48 files, 60 vectors); git diff --check PASS.

+++ src/lib/relaydesk/platform/MacLocalNetworkStatus.cpp
@@ -27,7 +27,7 @@ PermissionProbeEntry macLocalNetworkEntry(
     return {
         .kind = PermissionKind::MacLocalNetwork,
         .state = PermissionState::Denied,
-        .errorCode = static_cast<int>(PermissionErrorCode::MacLocalNetworkDenied),
+        .errorCode = PermissionErrorCode::MacLocalNetworkDenied,
         .canOpenSettings = true,
         .diagnostic = diagnostic,
     };
@@ -36,10 +36,9 @@ PermissionProbeEntry macLocalNetworkEntry(
   return {
       .kind = PermissionKind::MacLocalNetwork,
       .state = PermissionState::Unknown,
-      .errorCode = static_cast<int>(PermissionErrorCode::ProbeUnavailable),
+      .errorCode = PermissionErrorCode::ProbeUnavailable,
       .diagnostic = diagnostic,
   };
 }
 
 } // namespace deskflow::relaydesk
-
